+++ HISTORY.md
@@ -2,6 +2,9 @@ v0.1.5
 ======
 + use a fixed length of 4 in merge (gives much better merging)
 + add new annotate command which takes a gff
++ increase max_reads argument to svtyper to reduce false negatives
++ remove QCFail and duplicate reads from split and disc.bams. This is now
+  part of `smoove` but also fixed in `lumpy_filter` upstream.
 
 v0.1.4
 ======